design add to cart and cart items list component

+++ src/main/java/com/karam/gyw/model/CartModel.java
@@ -24,7 +24,7 @@ public class CartModel {
 		
 		@Id
 		@GeneratedValue(strategy =  GenerationType.IDENTITY)
-		private long product_id;
+		private int product_id;
 		
 		@Column(name = "product_name")
 		private String product_name;
@@ -32,11 +32,11 @@ public class CartModel {
 		@Column(name = "price")
 		private String price;
 
-		public long getProduct_id() {
+		public int getProduct_id() {
 			return product_id;
 		}
 
-		public void setProduct_id(long product_id) {
+		public void setProduct_id(int product_id) {
 			this.product_id = product_id;
 		}
