+++ src/main/java/com/karam/gyw/repository/CartRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.karam.gyw.model.CartModel;
 
 @Repository
-public interface CartRepository extends JpaRepository<CartModel, Long>{
+public interface CartRepository extends JpaRepository<CartModel, Integer>{
 
 }
